+++ backend/canisters/community/impl/src/queries/events_by_index.rs
@@ -10,7 +10,6 @@ fn events_by_index(args: Args) -> Response {
 
 fn events_by_index_impl(args: Args, state: &RuntimeState) -> Response {
     let caller = state.env.caller();
-    let now = state.env.now();
     let user_id = state.data.members.get(caller).map(|m| m.user_id);
 
     if !state.data.is_public && user_id.is_none() {
@@ -22,8 +21,8 @@ fn events_by_index_impl(args: Args, state: &RuntimeState) -> Response {
             user_id,
             args.thread_root_message_index,
             args.events,
+            args.latest_known_update,
             args.latest_client_event_index,
-            now,
         ) {
             EventsResult::Success(response) => Success(response),
             EventsResult::UserNotInGroup => UserNotInChannel,