+++ backend/canisters/user/impl/src/queries/events_window.rs
@@ -14,12 +14,14 @@ fn events_window_impl(args: Args, state: &RuntimeState) -> Response {
     if let Some(chat) = state.data.direct_chats.get(&args.user_id.into()) {
         let events_reader = chat.events.main_events_reader();
         let latest_event_index = events_reader.latest_event_index().unwrap();
+        let chat_last_updated = chat.last_updated();
 
-        if args.latest_client_event_index.map_or(false, |e| latest_event_index < e) {
+        if args.latest_known_update.map_or(false, |ts| chat_last_updated < ts)
+            || args.latest_client_event_index.map_or(false, |e| latest_event_index < e)
+        {
             return ReplicaNotUpToDate(latest_event_index);
         }
 
-        let now = state.env.now();
         let my_user_id = state.env.canister_id().into();
 
         let (events, expired_event_ranges) = EventOrExpiredRange::split(events_reader.window(
@@ -35,7 +37,8 @@ fn events_window_impl(args: Args, state: &RuntimeState) -> Response {
             expired_event_ranges,
             expired_message_ranges,
             latest_event_index,
-            timestamp: now,
+            chat_last_updated,
+            timestamp: chat_last_updated,
         })
     } else {
         ChatNotFound