+++ backend/integration_tests/src/client/user.rs
@@ -180,6 +180,7 @@ pub mod happy_path {
                 ascending,
                 max_messages,
                 max_events,
+                latest_known_update: None,
                 latest_client_event_index: None,
             },
         );
@@ -199,6 +200,7 @@ pub mod happy_path {
                 user_id,
                 thread_root_message_index: None,
                 events,
+                latest_known_update: None,
                 latest_client_event_index: None,
             },
         );